parser/requirements.txt
Flask
tvdb-api==1.9
tvnamer==2.3
omdb
python-dotenv
pyswagger
flask-envconfig